Move FolderAuthorizationWithMetadata test to foldermetadata package
This test exercises folder metadata behavior (stable UIDs vs hash-based
IDs) and belongs with the other folder metadata tests. Restored the
original per-variant RunGrafana calls so the "without metadata" variant
runs without the feature flag.

Made-with: Cursor
